+++ api/v1/zz_generated.deepcopy.go
@@ -197,6 +197,13 @@ func (in *LinstorClusterSpec) DeepCopyInto(out *LinstorClusterSpec) {
 		*out = new(corev1.NodeSelector)
 		(*in).DeepCopyInto(*out)
 	}
+	if in.Tolerations != nil {
+		in, out := &in.Tolerations, &out.Tolerations
+		*out = make([]corev1.Toleration, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
 	if in.Properties != nil {
 		in, out := &in.Properties, &out.Properties
 		*out = make([]LinstorControllerProperty, len(*in))