+++ internal/controller/patches.go
@@ -274,6 +274,28 @@ func SatelliteHostPathVolumeEnvPatch(hostPaths []string) ([]kusttypes.Patch, err
 	)
 }
 
+func TolerationsPatch(kind, name string, tolerations []corev1.Toleration) ([]kusttypes.Patch, error) {
+	patches, err := render(
+		cluster.Resources,
+		"patches/tolerations.yaml",
+		map[string]any{
+			"KIND":        kind,
+			"NAME":        name,
+			"TOLERATIONS": tolerations,
+		})
+	if err != nil {
+		return nil, err
+	}
+
+	for i := range patches {
+		patches[i].Target = &kusttypes.Selector{
+			ResId: resid.NewResIdKindOnly(kind, name),
+		}
+	}
+
+	return patches, nil
+}
+
 func ComponentPodTemplate(kind, name string, template json.RawMessage) ([]kusttypes.Patch, error) {
 	patches, err := render(
 		cluster.Resources,